+++ ErsatzTV.Application/Channels/ChannelViewModel.cs
@@ -19,6 +19,7 @@ public record ChannelViewModel(
     ChannelPlayoutSource PlayoutSource,
     ChannelPlayoutMode PlayoutMode,
     int? MirrorSourceChannelId,
+    TimeSpan? PlayoutOffset,
     StreamingMode StreamingMode,
     int? WatermarkId,
     int? FallbackFillerId,