+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `Generated`: default/existing behavior where channel must have its own playout
   - `Mirror`: channel will play content from the specified `Mirror Source Channel`'s playout
     - This allows the exact same content on different channels with different channel settings
+    - `Playout Offset` can be used to offset the times of scheduled playout items from the mirror source channel
+      - e.g. -2 hours will cause the mirror channel to play content 2 hours before the mirror source channel
 
 ### Fixed
 - Fix transcoding content with bt709/pc color metadata